make_flash_bem function moved to bem.py
